+++ src/Colors/Hsl/Decoders/StringColorDecoder.php
@@ -35,7 +35,8 @@ class StringColorDecoder extends AbstractDecoder implements DecoderInterface
      */
     public function decode(mixed $input): ColorInterface
     {
-        $pattern = '/^hsl\((?P<h>[0-9\.]+), ?(?P<s>[0-9\.]+%?), ?(?P<l>[0-9\.]+%?)\)$/i';
+        $pattern = '/^hsl\((?P<h>[0-9\.]+)((, ?)| )(?P<s>[0-9\.]+%?)((, ?)| )(?P<l>[0-9\.]+%?)\)$/i';
+
         if (preg_match($pattern, $input, $matches) != 1) {
             throw new InvalidArgumentException('Invalid hsl() color notation');
         }