+++ src/Colors/Hsv/Decoders/StringColorDecoder.php
@@ -35,7 +35,8 @@ class StringColorDecoder extends AbstractDecoder implements DecoderInterface
      */
     public function decode(mixed $input): ColorInterface
     {
-        $pattern = '/^hs(v|b)\((?P<h>[0-9\.]+), ?(?P<s>[0-9\.]+%?), ?(?P<v>[0-9\.]+%?)\)$/i';
+        $pattern = '/^hs(v|b)\((?P<h>[0-9\.]+)((, ?)| )(?P<s>[0-9\.]+%?)((, ?)| )(?P<v>[0-9\.]+%?)\)$/i';
+
         if (preg_match($pattern, $input, $matches) != 1) {
             throw new InvalidArgumentException('Invalid hsv() or hsb() color notation');
         }