+++ tests/Unit/Colors/Cmyk/Decoders/StringColorDecoderTest.php
@@ -5,38 +5,22 @@ declare(strict_types=1);
 namespace Intervention\Image\Tests\Unit\Colors\Cmyk\Decoders;
 
 use PHPUnit\Framework\Attributes\CoversClass;
-use Intervention\Image\Colors\Cmyk\Color;
 use Intervention\Image\Colors\Cmyk\Decoders\StringColorDecoder;
-use Intervention\Image\Exceptions\InvalidArgumentException;
 use Intervention\Image\Tests\BaseTestCase;
+use Intervention\Image\Tests\Providers\ColorDataProvider;
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 
 #[CoversClass(StringColorDecoder::class)]
 final class StringColorDecoderTest extends BaseTestCase
 {
-    public function testDecode(): void
+    /**
+     * @param $channelValues array<int>
+     */
+    #[DataProviderExternal(ColorDataProvider::class, 'cmykString')]
+    public function testDecode(mixed $input, array $channelValues): void
     {
         $decoder = new StringColorDecoder();
-        $result = $decoder->decode('cmyk(0,0,0,0)');
-        $this->assertInstanceOf(Color::class, $result);
-        $this->assertEquals([0, 0, 0, 0], $result->toArray());
-
-        $result = $decoder->decode('cmyk(0, 100, 100, 0)');
-        $this->assertInstanceOf(Color::class, $result);
-        $this->assertEquals([0, 100, 100, 0], $result->toArray());
-
-        $result = $decoder->decode('cmyk(0, 100, 100, 0)');
-        $this->assertInstanceOf(Color::class, $result);
-        $this->assertEquals([0, 100, 100, 0], $result->toArray());
-
-        $result = $decoder->decode('cmyk(0%, 100%, 100%, 0%)');
-        $this->assertInstanceOf(Color::class, $result);
-        $this->assertEquals([0, 100, 100, 0], $result->toArray());
-    }
-
-    public function testDecodeInvalid(): void
-    {
-        $decoder = new StringColorDecoder();
-        $this->expectException(InvalidArgumentException::class);
-        $decoder->decode(null);
+        $result = $decoder->decode($input[0]);
+        $this->assertEquals($channelValues, $result->toArray());
     }
 }