+++ tests/Unit/Colors/Hsl/Decoders/StringColorDecoderTest.php
@@ -4,12 +4,11 @@ declare(strict_types=1);
 
 namespace Intervention\Image\Tests\Unit\Colors\Hsl\Decoders;
 
-use Generator;
 use PHPUnit\Framework\Attributes\CoversClass;
-use Intervention\Image\Colors\Hsl\Color;
 use Intervention\Image\Colors\Hsl\Decoders\StringColorDecoder;
 use Intervention\Image\Tests\BaseTestCase;
-use PHPUnit\Framework\Attributes\DataProvider;
+use Intervention\Image\Tests\Providers\ColorDataProvider;
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 
 #[CoversClass(StringColorDecoder::class)]
 final class StringColorDecoderTest extends BaseTestCase
@@ -17,36 +16,11 @@ final class StringColorDecoderTest extends BaseTestCase
     /**
      * @param $channelValues array<int>
      */
-    #[DataProvider('decodeDataProvier')]
-    public function testDecode(string $input, string $classname, array $channelValues): void
+    #[DataProviderExternal(ColorDataProvider::class, 'hslString')]
+    public function testDecode(mixed $input, array $channelValues): void
     {
         $decoder = new StringColorDecoder();
-        $result = $decoder->decode($input);
-        $this->assertInstanceOf($classname, $result);
+        $result = $decoder->decode($input[0]);
         $this->assertEquals($channelValues, $result->toArray());
-    }
-
-    public static function decodeDataProvier(): Generator
-    {
-        yield [
-            'hsl(0,0,0)',
-            Color::class,
-            [0, 0, 0],
-        ];
-        yield [
-            'hsl(0, 100, 50)',
-            Color::class,
-            [0, 100, 50],
-        ];
-        yield [
-            'hsl(360, 100, 50)',
-            Color::class,
-            [360, 100, 50],
-        ];
-        yield [
-            'hsl(180, 100%, 50%)',
-            Color::class,
-            [180, 100, 50],
-        ];
     }
 }